+++ changes.md
@@ -8,6 +8,7 @@
 5. Let draw sub-items from TCanvas list of primitives like sub-pad or TLatex
 6. Optimize TGrapth drawing - web browsers have problem to draw 10000 pixels at the same time   
 7. Implement painter for TPolyMarker3D, works only with canvas
+8. Improve drawing and update of TMultiGraph 
 
 
 ## Changes in 4.0